Adding Task Killer and Task Start functionality
Of note, this works, but there is a lingering issue with respect to the system tray after tasks are killed via this method. If you try to kill explorer and restart it in the same script, the process hangs after explorer is killed due to the actual script being part of explorer's roll-up.